# Import: rewrite mesh importer with numpy