… addons/io_scene_gltf2/blender/imp/gltf2_blender_primitive.py