Rewrites the mesh import code to use numpy and foreach_set, as discussed in #996.

The first commit adds accessor decoding to numpy arrays. The existing functions are kept for compat with existing callers.

The second rewrites gltf2_blender_mesh.py. Code is pretty straightforward, just ask if there's anything you want to know about.

This is usually either good for performance, or neutral. Sometimes it is slower, but seems worth it on the balance.

|Model|master|this branch|
|:---|:---|:---|
|modelGroup.glb, Smooth Shading|110s|47s|
|MetalRoughSpheres|5.1|3.4s|
|AliciaSolid (iss967)|1.5s|0.5s|
|RubyRose (iss848)|2.1s|**2.4s**|

This code structure also makes it pretty easy to dedupe verts for #1069; PR to follow.