@scurest 2020-07-16
I realized I re-implemented the import shading handling wrong (in master, custom split normals is only set when "Use Normal Data" is on; here, I set it whenever we have normals). Let's block this until #1119 is resolved.