@scurest 2020-07-16
Nvm lol, I just went ahead and fixed it 'cause it was easy. For some reason I always thought the big difference between "Use Normal Data" and "Flat Shading"/"Smooth Shading" was how it set the poly's use_smooths, but it was actually that "Use Normal Data" creates custom normals and the others don't.

Since "Smooth Shading" now doesn't create split normals it's also even faster! modelGroup.glb is now a cool 35s :o