@scurest 2020-07-19
I pushed some more optimizations and RubyRose is no longer slower than master.