@julienduroure 2020-07-22
@Scurest 

Some comment from Blender developer team:

> Hi @Julien DUROURE (julien) it is advised to put imports inside function body to avoid inflating Blender startup time.
Many imports load very fast and have little impact, but numpy takes 0.14 seconds to load on my machine, and that is a significant portion of Blender's 2 second startup time.